Attempt at introducing jupyter lite
This is an in browser approach. What works now is detection of the notebook name and identifying it in a seamless manner. Ability to launch jupyter-lite also works.

What is not working:
- Scaling in the UI, iframe issues means I can't make this neatly fit, instead there is a fullscreen button
-Browser security won't allow me auto load a file into the browser for obvious security reasons.

For now the temp workaround is an autoload (that is still not working) and a create. So to this test click create, it downloads the notebook file and gives you instructions. if you are in fullscreen mode in the top nav bar there is an up arrow to upload, select the file and it should load successfully.

So this shows it can be done, there just needs to be some better UX applied (easy to do) but a smarter approach to autoloading the notebook that respects security is also needed